Basic structure to check if we can access Weather API without problems
For a while, we will only be displaying the temperature,
and it's only in Celsius.

The user can enter a specific location (city, state, country).

I'm aware that API keys shouldn't be exposed like I'm doing
here. But, this project is part of a course, and only later
during backend lessons I'll learn ways to securely deal with
these topics.