Reduce the amount of code
Change 'for... in' loop in the index.js file for
Object.values() method.

The new method exists inside a new function called
displayData(), wich helps avoid writing the same
block of code again when dealing with forms.

The result is cleaner code outside the factory
function in index.js file.