Get and pass forecast data to the forecast container
Make a copy of the functions you have been using
to get weather data and adjust to the forecast
context.

Make a copy of the async function getWeather() and
change it to the forecast context.

Create and call a function in the index.js file that
is capable of integrate forecast container's DOM
elements and its data.