Improve performance - current weather
Previously, to get the current weather for a certain
location, there were an unnecessary amount of requests
to the API.

The excessive API calls were related to a loop,
where the requisition and the access to a specific
information were all executed together.

To fix that, we split 'accessWeatherAPI' into two
modules and only one of them is called inside the
loop responsible for the requisitions, avoiding
the problem.