Improve performance - weather forecast
Previously, to get the forecast weather for a given
location, there were an unnecessary amount of requests
to the API.

The excessive API calls were related to a loop,
where the requisition and the access to a specific
information were all executed together.

To fix that, we split 'accessForecastAPI' into two
modules and only one of them is called inside the
loop responsible for the requisitions, avoiding
the problem.